- einige Änderungen - Formular einfügen und auslesen Version 1 (funktioniert aber noch nicht - irgendwas mit dem TreeSet...)

+++ Stuffbox/src/com/stuffbox/model/DataSourceItem.java
@@ -92,7 +92,7 @@ public class DataSourceItem {
 		whereStatement.append(" ");
 		whereStatement.append(DatabaseHandler.SQL_AND);
 		
-    	ArrayList<Integer> selectFeatureIds = new ArrayList<Integer>();
+    	ArrayList<Long> selectFeatureIds = new ArrayList<Long>();
     	for(Feature feature : features){
     		selectFeatureIds.add(feature.getId());
     	}
